backend/requirements.txt
asgiref==3.5.2
Django==4.1.4
djangorestframework==3.14.0
keras==2.11.0
numpy==1.24.1
Pillow==9.3.0
pytz==2022.6
sklearn==0.0.post1
sqlparse==0.4.3
